octoprint_compat: add webcam_rotation parameter
octoprint_compat: add webcam_rotation parameter

Add `webcam_rotation` parameter to the `[octoprint_compat]` settings to allow specifying a rotation in degrees (0, 90, 180, 270) to match the `[webcam]` `rotation` parameter. Since a horizontal flip and a vertical flip combined are equivalent to a 180-degree rotation, this is implemented by setting `flipH` and `flipV` to true for a 180 degree rotation and setting those parameter plus `rotate90` to true for a 270-degree rotation.

Signed-off-by: Zan Hecht <github@zansstuff.com>