Explain stream types
Add a list of allowed stream types. I thought this was valuable because the stream types are not explained in this document. I had to look in the fluidd codebase to find a list of valid values. The valid values are not the same as the options shown in the fluidd webUI.

https://github.com/fluidd-core/fluidd/blob/999ccddd1a39d6037ec6a8f34b689b0216ea592d/src/components/settings/cameras/CameraConfigDialog.vue#L86