update_manager: add report_anomalies option
Anomalies do not prevent updates, they are intended to alert
users of potential issues detected within an extension for
update.  Anomalies are always logged, however it is not
necessary for them to be reported to frontends.  This option
gives users the ability to silence anomalies on a per extension
basis.

Signed-off-by:  Eric Callahan <arksine.code@gmail.com>