job_state: refactor print start detection
Any transition to from an idle state to any state other
than "standby" should emit a started event.  This should
nearly always be a "printing" state, however it is possible
that Klipper's API server will not have a chance to push
a  "printing" status update if the print is immediately
paused, cancelled, or aborted due to an error.

Signed-off-by:  Eric Callahan <arksine.code@gmail.com>