metadata: improve handling of file encoding
The size of the file is not necessarily equal to the number of characters
read due to variable length utf8 encoding.  Read gcode files in binary mode then encode to a string.

When determining the "gcode start byte" and "gcode end byte" use
the length of the decoded string to determine the correct value.

Signed-off-by:  Eric Callahan <arksine.code@gmail.com>